Add Traefik support to Workstation containers
Each container entry can specify its own 'traefikCreator' which is a function that is called by the host Traefik config and returns an attribute set of the container's desired Traefik configuration. This was added so that container-specific configuration can stay out of the host files, while still allowing the host to serve container web traffic. This allows LAN devices to connect to the Workstation and establish a connection with a container.