Make Traefik containers use domain instead of path
As much as I wanted to avoid using a domain name, it seems that Reverse Proxy paths cause a lot of website pathing issues that don't occur when you use domain names. In this case, Selkies-Gstreamer couldn't load files by relative path because Traefik with its PathPrefix was causing relative paths to be treated as absolute paths. Every forum post I read said to basically use a domain name instead, so here we are..