Fix PROMPT_COMMAND ble.sh slowness
Earlier, PROMPT_COMMAND was set so that history was shared. Later on ble.sh was added to the equation and it comes with its own history sharing features which causes blatant slowness issues in containers when the PROMPT_COMMAND hack is also set. To fix this, the PROMPT_COMMAND hack is now ignored if ble.sh is being used.