Switch from bash to zsh
bash doesn't have the same features as zsh. There is only a single option to compete which you do by installing ble.sh. Ble.sh's performance is absolutely abyssmal as it literally does a hundred different calls every time you press tab (I tried disabling each option, however this seems to be a problem with the core of ble.sh). If that wasn't bad enough, ble.sh's slowness is amplified to unusuable when used inside of containers. This is unacceptable for any shell, so I've made the decision to enforce zsh usage instead so that the same featureset can be used without any hits to performance.