Fix CPU performance with Selkies Gstreamer
Systemd CPUWeight is now applied for Selkies Gstreamer services (500, so the stream and desktop continue to be performant under heavy CPU load) and Workstation containers (less than 100 so that the host always gets the most CPU time). This makes containerization with Selkies Gstreamer instances A LOT smoother, less lags and freeze ups.